job-recommender/requirements.txt
streamlit==1.48.1
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
tqdm==4.66.4

# LangChain + Gemini
langchain==0.3.7
langchain-community==0.3.7
google-generativeai==0.8.5
langchain-google-genai==2.0.10

# FAISS only (no ChromaDB)
faiss-cpu==1.12.0
